pyproject.toml
[project]
name = "nowcasting"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-bcb>=0.2",
    "sidrapy>=0.1",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "statsmodels>=0.14",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "jupyterlab>=4.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
markers = [
    "slow: integration tests that hit real APIs (deselect with '-m \"not slow\"')",
]
